GitHub Action: Split .docx and .pdf building
Both on push, but we split them since the pandoc conversion is actually
way way faster then the native pdflatex build (almost a factor 10...).

So we want to take full advantage of pandoc's speed and get right away
the artifact, irrespective of what happens to the CI-intended PDF build.

Let's see...