PARTIALLY LEAVING EXPORT_FIG → RESURRECTING FIG2SVG
- All the histograms are made raster, which is unacceptable (I don't care the reason, just no, nope)

- The GitHub runner hangs on, for that export_fig invokes deprecated features on -nodisp environment

So now we mix every exporting library we are have at hand: export_fig were it excels, exportgraphics
for basically all histograms except the one made with barplots, which goes really crazy and displays
nonsensical diagonals on the bars (WTF MATLAB), so we end up resurrecting SVG files with fig2svg. :O

Of course that unique SVG needs to be converted to PDF before being included in the LaTeX, which we
do with InkScape. Since we don't want, FOR ANY REASON, to overwrite that PDF with an accidental run
of the script I explicitly add < *.svg > to the .gitignore. Blame me if you dare. Jokes apart, that
would also turn useful for all the failed exports that happen IN WHATEVER OTHER CASE we run fig2svg

Over and out.