GitHub Actions: chmod apparently does not work on the runner
From bash manual:

If a command is found but is not executable, the return status is 126.
> https://www.gnu.org/software/bash/manual/html_node/Exit-Status.html

From the damn action log:

./pdf2archive: Is a directory
Error: Process completed with exit code 126.

So, despite the message being "is a directory" the actual problem is
that the file is not executable. Fuck.

NOTE) The thing must be very subtle, since previously we got
      -rwxr-xr-x 1 runner docker  16K Jun 15 00:50 pdf2archive
      from a ls -lh command. WTF.

      > let's try with chmode a+rx